finished moving msg to chat
